mcp tests refactoring Assisted-by: Claude Sonnet 4.5 (via Claude Code)
